+++ lib/thread_pool.cpp
@@ -1,9 +1,9 @@
 #include "thread_pool.hpp"
 #include <functional>
 
-thread_pool::thread_pool(int threads) : in_destructor_(false), waiting_on_(0) {
+thread_pool::thread_pool(size_t threads) : in_destructor_(false), waiting_on_(0) {
     threads_.reserve(threads);
-    for (int i = 0; i < threads; i++) threads_.emplace_back(std::bind(&thread_pool::enter_thread, this));
+    for (size_t i = 0; i < threads; i++) threads_.emplace_back(std::bind(&thread_pool::enter_thread, this));
 }
 
 thread_pool::~thread_pool() {
